+++ lib/servo/src/inst/inst_transport_interface.h
@@ -27,10 +27,10 @@ class InstTransportInterface {
 
   /**
    * @brief 处理数据
-   * @param dt 时间间隔(微秒)
+   * @param dt 时间间隔(秒)
    * @return 错误码
    */
-  virtual Error Process(const uint32_t dt) = 0;
+  virtual Error Process(const float dt) = 0;
 
   /**
    * @brief 发送数据
